requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-gitlab==4.4.0
openai==0.28.0
tiktoken==0.5.2

# Vector stores
qdrant-client==1.7.0

# Utilities
pydantic==2.5.3
python-dotenv==1.0.0
pyyaml==6.0.1
aiofiles==23.2.1
httpx>=0.27.0
numpy

# Logging and monitoring
python-json-logger==2.0.7
prometheus-client==0.19.0

# Development and testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
black==23.12.1
flake8==7.0.0
mypy==1.8.0